refactor(trasaction): create recipient component
