+++ src/modules/transactions/components/dialog/create/recipient/index.tsx
@@ -1,0 +1,9 @@
+import Form from './Form';
+import Item from './Item';
+import List from './List';
+
+export const Recipient = {
+  List,
+  Item,
+  Form,
+};
